requirements.txt
streamlit>=1.28.0
openai>=1.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
asyncio-mqtt>=0.11.0
aiohttp>=3.8.0
pydantic>=2.0.0
tiktoken>=0.5.0
httpx[socks]>=0.24.0
socksio>=1.0.0